extract i18next init to separate file
To avoid cluttering app.js more than necessary, I am extracting the logic to initialize i18next to a new file called i18nextInit.ts, which just performs the same initialization logic and exports the initialized instance of i18next.

In tsconfig.json, `resolveJsonModule` was set `true` so that the JSON translation files can be imported in TypeScript.
