+++ tsconfig.json
@@ -5,6 +5,7 @@
     "module": "es6",
     "target": "es5",
     "esModuleInterop": true,
+    "resolveJsonModule": true,
     "jsx": "react",
     "allowJs": true,
     "moduleResolution": "node"
